+++ fastapi/notas_fiscais/app/model/danfe_model.py
@@ -1,7 +1,8 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
-from sqlalchemy import DateTime, Float, ForeignKey, Integer, String
+from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database.connection import Base
@@ -17,7 +18,7 @@ class DanfeModel(Base):
     id_danfe: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     cd_contribuinte: Mapped[str] = mapped_column(String(20), ForeignKey("NOTA_FISCAL.CONTRIBUINTE.cd_contribuinte"), nullable=False)
     numero: Mapped[str] = mapped_column(String(15))
-    valor_total: Mapped[float] = mapped_column(Float)
+    valor_total: Mapped[Decimal] = mapped_column(Numeric(12,2))
     data_emissao: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
 
     contribuinte: Mapped["ContribuinteModel"] = relationship(
